+++ src/include/kenwood/TS850.h
@@ -15,7 +15,7 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// aunsigned long int with this program.  If not, see <http://www.gnu.org/licenses/>.
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // ----------------------------------------------------------------------------
 
 #ifndef _TS850_H
@@ -40,15 +40,15 @@ public:
 
 	bool check();
 
-	unsigned long int get_vfoA();
-	void set_vfoA(unsigned long int);
+	unsigned long long get_vfoA();
+	void set_vfoA(unsigned long long);
 	int  get_modeA();
 	void set_modeA(int val);
 	void set_bwA(int val);
 	int  get_bwA();
 
-	unsigned long int get_vfoB();
-	void set_vfoB(unsigned long int);
+	unsigned long long get_vfoB();
+	void set_vfoB(unsigned long long);
 	int  get_modeB();
 	void set_modeB(int val);
 	void set_bwB(int val);
