+++ src/include/yaesu/FTdx10.h
@@ -15,7 +15,7 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// aunsigned long int with this program.  If not, see <http://www.gnu.org/licenses/>.
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // ----------------------------------------------------------------------------
 
 #ifndef _FTdx10_H
@@ -35,11 +35,11 @@ public:
 
 	virtual bool check();
 
-	virtual unsigned long int get_vfoA();
-	virtual void set_vfoA(unsigned long int);
+	virtual unsigned long long get_vfoA();
+	virtual void set_vfoA(unsigned long long);
 
-	virtual unsigned long int get_vfoB();
-	virtual void set_vfoB(unsigned long int);
+	virtual unsigned long long get_vfoB();
+	virtual void set_vfoB(unsigned long long);
 
 	virtual int get_vfoAorB();
