+++ src/rigs/elad/FDMDUO.cxx
@@ -15,7 +15,7 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// aunsigned long int with this program.  If not, see <http://www.gnu.org/licenses/>.
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // ----------------------------------------------------------------------------
 
 #include "elad/FDMDUO.h"
@@ -130,7 +130,7 @@ RIG_FDMDUO::RIG_FDMDUO() {
 	serial_dtrptt = false;
 	B.imode = A.imode = 1;
 	B.iBW = A.iBW = 0x8A03;
-	B.freq = A.freq = 14070000;
+	B.freq = A.freq = 14070000ULL;
 
 	can_change_alt_vfo = true;
 
@@ -849,7 +849,7 @@ int RIG_FDMDUO::get_split()
 	return 0;
 }
 
-unsigned long int RIG_FDMDUO::get_vfoA ()
+unsigned long long RIG_FDMDUO::get_vfoA ()
 {
 	cmd = "FA;";
 
@@ -861,7 +861,7 @@ unsigned long int RIG_FDMDUO::get_vfoA ()
 
 	size_t p = replystr.rfind("FA");
 	if (p != std::string::npos && (p + 12 < replystr.length())) {
-		int f = 0;
+		unsigned long long f = 0;
 		for (size_t n = 2; n < 13; n++)
 			f = f*10 + replystr[p+n] - '0';
 		A.freq = f;
@@ -869,7 +869,7 @@ unsigned long int RIG_FDMDUO::get_vfoA ()
 	return A.freq;
 }
 
-void RIG_FDMDUO::set_vfoA (unsigned long int freq)
+void RIG_FDMDUO::set_vfoA (unsigned long long freq)
 {
 	A.freq = freq;
 	cmd = "FA00000000000;";
@@ -881,7 +881,7 @@ void RIG_FDMDUO::set_vfoA (unsigned long int freq)
 	showresp(WARN, ASC, "set vfo A", cmd, "");
 }
 
-unsigned long int RIG_FDMDUO::get_vfoB ()
+unsigned long long RIG_FDMDUO::get_vfoB ()
 {
 	cmd = "FB;";
 
@@ -893,7 +893,7 @@ unsigned long int RIG_FDMDUO::get_vfoB ()
 
 	size_t p = replystr.rfind("FB");
 	if (p != std::string::npos && (p + 12 < replystr.length())) {
-		int f = 0;
+		unsigned long long f = 0;
 		for (size_t n = 2; n < 13; n++)
 			f = f*10 + replystr[p+n] - '0';
 		B.freq = f;
@@ -901,7 +901,7 @@ unsigned long int RIG_FDMDUO::get_vfoB ()
 	return B.freq;
 }
 
-void RIG_FDMDUO::set_vfoB (unsigned long int freq)
+void RIG_FDMDUO::set_vfoB (unsigned long long freq)
 {
 	B.freq = freq;
 	cmd = "FB00000000000;";
