+++ src/rigs/elecraft/K3.cxx
@@ -15,7 +15,7 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// aunsigned long int with this program.  If not, see <http://www.gnu.org/licenses/>.
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // ----------------------------------------------------------------------------
 
 #include "elecraft/K3.h"
@@ -81,7 +81,7 @@ RIG_K3::RIG_K3() {
 	serial_rtsptt = false;
 	serial_dtrptt = false;
 
-	def_freq = freqA = freqB = 14070000;
+	def_freq = freqA = freqB = 14070000ULL;
 	def_mode = modeA = modeB = 1;
 	def_bw = bwA = bwB = 34;
 
@@ -202,7 +202,7 @@ bool RIG_K3::check ()
 	return true;
 }
 
-unsigned long int RIG_K3::get_vfoA ()
+unsigned long long RIG_K3::get_vfoA ()
 {
 	cmd = "FA;";
 	get_trace(1, "get vfoA");
@@ -213,14 +213,14 @@ unsigned long int RIG_K3::get_vfoA ()
 	size_t p = replystr.rfind("FA");
 	if (p == std::string::npos) return freqA;
 
-	unsigned long int f = 0;
+	unsigned long long f = 0;
 	for (size_t n = 2; n < 13; n++)
 		f = f*10 + replystr[p + n] - '0';
 	freqA = f;
 	return freqA;
 }
 
-void RIG_K3::set_vfoA (unsigned long int freq)
+void RIG_K3::set_vfoA (unsigned long long freq)
 {
 	freqA = freq;
 	cmd = "FA00000000000;";
@@ -233,7 +233,7 @@ void RIG_K3::set_vfoA (unsigned long int freq)
 	sett("");
 }
 
-unsigned long int RIG_K3::get_vfoB ()
+unsigned long long RIG_K3::get_vfoB ()
 {
 	cmd = "FB;";
 	get_trace(1, "get vfoB");
@@ -244,14 +244,14 @@ unsigned long int RIG_K3::get_vfoB ()
 	size_t p = replystr.rfind("FB");
 	if (p == std::string::npos) return freqB;
 
-	unsigned long int f = 0;
+	unsigned long long f = 0;
 	for (size_t n = 2; n < 13; n++)
 		f = f*10 + replystr[p + n] - '0';
 	freqB = f;
 	return freqB;
 }
 
-void RIG_K3::set_vfoB (unsigned long int freq)
+void RIG_K3::set_vfoB (unsigned long long freq)
 {
 	freqB = freq;
 	cmd = "FB00000000000;";
