+++ src/rigs/elecraft/K4.cxx
@@ -15,7 +15,7 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// aunsigned long int with this program.  If not, see <http://www.gnu.org/licenses/>.
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // ----------------------------------------------------------------------------
 
 #include "elecraft/K4.h"
@@ -94,7 +94,7 @@ RIG_K4::RIG_K4() {
 	serial_rtsptt = false;
 	serial_dtrptt = false;
 
-	def_freq = freqA = freqB = 14070000;
+	def_freq = freqA = freqB = 14070000ULL;
 	def_mode = modeA = modeB = 1;
 	def_bw = bwA = bwB = 34;
 
@@ -219,7 +219,7 @@ bool RIG_K4::check ()
 	return true;
 }
 
-unsigned long int RIG_K4::get_vfoA ()
+unsigned long long RIG_K4::get_vfoA ()
 {
 	cmd = "FA;";
 	get_trace(1, "get vfoA");
@@ -229,14 +229,14 @@ unsigned long int RIG_K4::get_vfoA ()
 	size_t p = replystr.rfind("FA");
 	if (p == std::string::npos) return freqA;
 
-	unsigned long int f = 0;
+	unsigned long long f = 0;
 	for (size_t n = p + 2; n < replystr.length() - 1; n++)
 		f = f*10 + replystr[n] - '0';
 	freqA = f;
 	return freqA;
 }
 
-void RIG_K4::set_vfoA (unsigned long int freq)
+void RIG_K4::set_vfoA (unsigned long long freq)
 {
 	freqA = freq;
 	cmd = "FA00000000000;";
@@ -249,7 +249,7 @@ void RIG_K4::set_vfoA (unsigned long int freq)
 	sett("");
 }
 
-unsigned long int RIG_K4::get_vfoB ()
+unsigned long long RIG_K4::get_vfoB ()
 {
 	cmd = "FB;";
 	get_trace(1, "get vfoB");
@@ -259,14 +259,14 @@ unsigned long int RIG_K4::get_vfoB ()
 	size_t p = replystr.rfind("FB");
 	if (p == std::string::npos) return freqB;
 
-	unsigned long int f = 0;
+	unsigned long long f = 0;
 	for (size_t n = p + 2; n < replystr.length() - 1; n++)
 		f = f*10 + replystr[n] - '0';
 	freqB = f;
 	return freqB;
 }
 
-void RIG_K4::set_vfoB (unsigned long int freq)
+void RIG_K4::set_vfoB (unsigned long long freq)
 {
 	freqB = freq;
 	cmd = "FB00000000000;";
